add workernode: job distribution

+++ trunk/src/quads/kdtree/kdtree_io.c
@@ -464,7 +464,7 @@ int kdtree_write(FILE* fout, kdtree_t* kdtree) {
 }
 
 kdtree_t* kdtree_read(FILE* fin, int use_mmap,
-					  void** mmapped, int* mmapped_size) {
+					  void** mmapped, size_t* mmapped_size) {
     unsigned char ver, rs, is, kdns;
     unsigned int endian, ns;
     unsigned int N, D, nnodes;
